LayoutTests/fast/images/text-recognition/image-overlay-creation-is-idempotent-expected.txt

PASS internals.shadowRoot(image).childElementCount is 2
PASS successfullyParsed is true

TEST COMPLETE
